Fix bad command bug & restructure abstract command logic
There existed a bug where if a lot of long-running commands of the same type,
the last running command in the queue would change the MessageReceivedEvent
for the command, causing the results of the previous command to be sent to
the channel of the last running command instead. This fixes said bug by
adding a parameter to Command#onCommand() and doing replies and such there.