Migrate from JUnit to TestNG
We want to run the DatabaseInitializationTests class first, so it creates tables in CI before other database tests run (as of this commit, there aren't any, but some will be added later.) JUnit doesn't seem to support this in a clean way, so we'll just use TestNG instead.

Also adds some documentation on how to create a test configuration in IntelliJ Ultimate.