Welcome. Welcome to Java 17.
You have chosen, or been chosen, to relocate to one of our finest
remaining urban centers. I thought so much of Java 17 that I elected
to establish my Administration here, in the Citadel so thoughtfully
provided by Our Benefactors. I have been proud to call Java 17 my
home. And so, whether you are here to stay, or passing through on your
way to parts unknown, welcome to Java 17. It's safer here.

This commit upgrades Gradle to 7.3 and sets CI to build for Java 17.