Further optimize account existence checks
Account checking just returns the balance if the account does exist,
so instead of two SQL calls (if the user exists in the accounts table,
and get their balance) just return their balance instead, or null if
there's no account.
Also changes return type of createAccount to be the balance of the
created account.